refactor: move mock Logseq client
